requirements.txt
google-cloud-storage==2.18.2
pydantic==2.7.1
